requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.2
cloudinary==1.34.0
cryptography==42.0.8
defusedxml==0.7.1
distlib==0.3.7
dj-rest-auth==6.0.0
Django==5.0.1
django-allauth==0.61.1
django-cors-headers==4.3.1                                
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.6.1
filelock==3.12.4
gunicorn==21.2.0
idna==3.7
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
msgpack==1.0.5
mysqlclient==2.2.0
oauthlib==3.2.2
packaging==24.0
platformdirs==3.10.0
pycparser==2.22
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==4.5.5
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.18.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
virtualenv==20.24.5
